% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_update.R
\name{bayes_update}
\alias{bayes_update}
\title{Bayesian Updating}
\usage{
bayes_update(pretest_probability, lr_positive, lr_negative)
}
\arguments{
\item{pretest_probability}{Numeric value between 0 and 1 representing
the pre-test probability.}

\item{lr_positive}{Positive likelihood ratio.}

\item{lr_negative}{Negative likelihood ratio.}
}
\value{
An object of class "bayes_update" containing:
\describe{
\item{pretest_probability}{Original pre-test probability.}
\item{post_positive}{Post-test probability following a positive test.}
\item{post_negative}{Post-test probability following a negative test.}
}
}
\description{
Calculate post-test probabilities from a pre-test probability and
positive/negative likelihood ratios.
}
