% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eba_interpretation.R
\name{eba_interpretation}
\alias{eba_interpretation}
\title{Evidence-Based Assessment Interpretation}
\usage{
eba_interpretation(
  post_probability,
  wait_threshold = 0.1,
  treat_threshold = 0.7
)
}
\arguments{
\item{post_probability}{Numeric value between 0 and 1.}

\item{wait_threshold}{Threshold below which the condition is considered
unlikely. Default = 0.10.}

\item{treat_threshold}{Threshold above which the condition is considered
likely. Default = 0.70.}
}
\value{
An object of class "eba_interpretation".
}
\description{
Interpret a post-test probability using evidence-based assessment
decision thresholds.
}
