% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eba_stats.R
\name{eba_stats}
\alias{eba_stats}
\title{Evidence-Based Assessment Statistics}
\usage{
eba_stats(tp, fp, fn, tn)
}
\arguments{
\item{tp}{Number of true positives.}

\item{fp}{Number of false positives.}

\item{fn}{Number of false negatives.}

\item{tn}{Number of true negatives.}
}
\value{
An object of class \code{"eba_stats"}.
}
\description{
Calculate evidence-based assessment statistics from a 2 × 2 contingency table.
}
\examples{
eba_stats(
  tp = 50,
  fp = 10,
  fn = 5,
  tn = 100
)

}
