% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagan_nomogram.R
\name{fagan_nomogram}
\alias{fagan_nomogram}
\title{Fagan Nomogram}
\usage{
fagan_nomogram(pretest_probability, lr_positive, lr_negative)
}
\arguments{
\item{pretest_probability}{Numeric value between 0 and 1.}

\item{lr_positive}{Positive likelihood ratio.}

\item{lr_negative}{Negative likelihood ratio.}
}
\value{
Invisibly returns the calculated probabilities.
}
\description{
Plot pre-test and post-test probabilities using likelihood ratios.
}
